Configure autoheal (#74)
* Fix autoheal config

* Rename .java to .kt

* Fix after rebase

* Fix after rebase

+++ pdanet/src/main/java/net/artux/pdanetwork/service/user/ban/BanService.java
@@ -14,9 +14,9 @@ public interface BanService {
 
     boolean setChatBan(UUID userId);
 
-    BanDto applySystemBan(UUID userId, String reason, String message, int secs);
+    BanDto applySystemBan(UUID userId, String reason, String message, long secs);
 
-    BanDto applyBan(UUID userId, String reason, String message, int secs);
+    BanDto applyBan(UUID userId, String reason, String message, long secs);
 
     BanDto getCurrentBan(UUID userId);
